feat(mespapiers): Add bypass to the SelectPaperVersion modal
To improve, the general idea is that for articles with
several versions, there is not systematically the
SelectPaperVersion modal which appears to ask the
user to confirm the version.
But to be able to activate it only for certain papers.
